Fix DPI scaling for status bar. Add DHT node count.

+++ src/UI/StatusBar.hpp
@@ -2,6 +2,8 @@
 
 #include "../stdafx.h"
 
+#include <stdint.h>
+
 namespace UI
 {
 class StatusBar
@@ -9,7 +11,7 @@ class StatusBar
 public:
     HWND Create(HWND hWndParent, RECT rc);
 
-    void SetDhtNodes(int nodes);
+    void SetDhtNodes(int64_t nodes);
     void SetTorrentCount(int total, int selected);
     void SetTransferRates(int dl, int ul);
